// src/app/modules/product/product.interface.ts
export interface TEyeglasses {
  productName: string;
  productPrice: number;
  productQuantity: number;
  frameMaterial: 'metal' | 'plastic' | 'acetate';
  frameShape: 'rectangular' | 'round' | 'cat-eye';
  lensType: string;
  brand: string;
  gender: 'men' | 'women' | 'unisex';
  productImage: string;
  color: string;
  userEmail: string;
}